pyproject.toml
[tool.poetry]
name = "graphql-tools"
version = "0.1.0"
description = "GraphQL Tools written in Python"
authors = ["Aditya Karnam <akarnam37@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "graphql_tools"}]

[tool.poetry.dependencies]
python = "^3.11"
graphql-core = "^3.2.3"
faker = "^18.3.1"
pydantic = "^1.10.7"
black = "^23.1.0"
pytest = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
